Fix location of benchmarks sources
The location of the benchmark sources changed when we improved our
integration testing infrastructure, but the change wasn't reflected in
the benchmarks.

Because the location of the sources can no longer be known in advance,
we need to use the utilities defined in the test suite of Bloop.
Unfortunately, this project doesn't depend on it.

We use reflection to call these utilities. We just need to be careful to
call them at most once per fork.