refactor: don't include the sbt-bloop-build
There is no real reason to do this (from what I can tell) since
everything that is needed to actually run the benchmarks are set
up on the Bloop side. I've fairly sure that this use case is just
like the other use case in Bloop where we want to dogfood sbt-bloop
so we do all sorts of stuff just to include it. By removing this it
allows us to fully just get rid of all the publishing of the naked/shaded
meta meta build stuff